refactor(requests): reading a request and writing one stop sharing a file
`inspection-request.service.ts` sat at 500 lines against a 501 cap and is on
booking-deposit (#20)'s path. Pure move plus one deduplication, described
below; the 138 tests across tests/unit/{inspections,usage,bookings} pass
untouched.

The seam is read vs write, and the two share nothing but table names.
Writing is quota, template-ownership validation, inserts and the
`inspection_people` mirror. Reading is one projection — parent plus children
plus template names — and its correctness rests on an invariant that has
nothing to do with writing:

CLIENT NAME COMES FROM `inspection_people`, NEVER FROM
`inspections.client_name`. That column survives GDPR erasure as a stale
denormalized cache, so reading it leaks an erased subject's name. The Task 9c
comment saying so was written twice, above two verbatim copies of a
four-table LEFT JOIN — once in `list`, once in `get`. In
`inspection-request/request-read.ts` it is `selectSubInspections`, said once,
narrowed by the caller's condition. THAT IS THE ONE NON-MOVE IN THIS COMMIT:
same joins, same order (role filter first, so the join does not fan out over
every role), same projection, same `where` shape with the caller's predicate
in the position the caller's predicate was in.

`ListFilter`, `SubInspectionRow` and `RequestRow` go with it; only
`ListFilter` comes back, as a type import for the delegating signature.
`shapeRequest` is module-private, not exported — nothing outside may build
the API shape by hand, and the dead-code gate agrees.

`list` and `get` stay as methods. `get` in particular is reloaded through by
`create`, `addSubInspection`, `update` and `getByInspectionId`, so making it
a free function would have rewritten four call sites inside the same file for
no gain.

Baseline ENTRY removed, not tightened. 345 lines is under the ordinary
400-line rule; re-baselining at 345 would have rebuilt the wall.
